+++ nipype/interfaces/utility/__init__.py
@@ -6,8 +6,6 @@ Package contains interfaces for using existing functionality in other packages
 
 Requires Packages to be installed
 """
-from __future__ import print_function, division, unicode_literals, absolute_import
-__docformat__ = 'restructuredtext'
 
 from .base import (IdentityInterface, Rename, Select, Split, Merge,
                    AssertEqual)
