+++ nipype/interfaces/utility/tests/test_wrappers.py
@@ -1,38 +1,12 @@
 # -*- coding: utf-8 -*-
-from __future__ import print_function, unicode_literals
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
+from __future__ import print_function, unicode_literals
 import os
 import pytest
 
 from nipype.interfaces import utility
 import nipype.pipeline.engine as pe
-
-
-def test_rename(tmpdir):
-    os.chdir(str(tmpdir))
-
-    # Test very simple rename
-    _ = open("file.txt", "w").close()
-    rn = utility.Rename(in_file="file.txt", format_string="test_file1.txt")
-    res = rn.run()
-    outfile = str(tmpdir.join("test_file1.txt"))
-    assert res.outputs.out_file == outfile
-    assert os.path.exists(outfile)
-
-    # Now a string-formatting version
-    rn = utility.Rename(in_file="file.txt", format_string="%(field1)s_file%(field2)d", keep_ext=True)
-    # Test .input field creation
-    assert hasattr(rn.inputs, "field1")
-    assert hasattr(rn.inputs, "field2")
-
-    # Set the inputs
-    rn.inputs.field1 = "test"
-    rn.inputs.field2 = 2
-    res = rn.run()
-    outfile = str(tmpdir.join("test_file2.txt"))
-    assert res.outputs.out_file == outfile
-    assert os.path.exists(outfile)
 
 
 def test_function(tmpdir):
@@ -87,48 +61,6 @@ def test_function_with_imports(tmpdir):
     print(node.inputs.function_str)
     node.inputs.size = 10
     node.run()
-
-
-@pytest.mark.parametrize("args, expected", [
-        ({}                ,  ([0], [1,2,3])),
-        ({"squeeze" : True},  (0  , [1,2,3]))
-        ])
-def test_split(tmpdir, args, expected):
-    os.chdir(str(tmpdir))
-
-    node = pe.Node(utility.Split(inlist=list(range(4)),
-                                 splits=[1, 3],
-                                 **args),
-                   name='split_squeeze')
-    res = node.run()
-    assert res.outputs.out1 == expected[0]
-    assert res.outputs.out2 == expected[1]
-
-
-def test_csvReader(tmpdir):
-    header = "files,labels,erosion\n"
-    lines = ["foo,hello,300.1\n",
-             "bar,world,5\n",
-             "baz,goodbye,0.3\n"]
-    for x in range(2):
-        name = str(tmpdir.join("testfile.csv"))
-        with open(name, 'w') as fid:
-            reader = utility.CSVReader()
-            if x % 2 == 0:
-                fid.write(header)
-                reader.inputs.header = True
-            fid.writelines(lines)
-            fid.flush()
-            reader.inputs.in_file = name
-            out = reader.run()
-            if x % 2 == 0:
-                assert out.outputs.files == ['foo', 'bar', 'baz']
-                assert out.outputs.labels == ['hello', 'world', 'goodbye']
-                assert out.outputs.erosion == ['300.1', '5', '0.3']
-            else:
-                assert out.outputs.column_0 == ['foo', 'bar', 'baz']
-                assert out.outputs.column_1 == ['hello', 'world', 'goodbye']
-                assert out.outputs.column_2 == ['300.1', '5', '0.3']
 
 
 def test_aux_connect_function(tmpdir):
